apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "citypulse-api"
version = "0.1.0"
description = "FastAPI backend scaffold for CityPulse."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1,<2.0.0",
  "aiosqlite>=0.21.0,<1.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "boto3>=1.37.0,<2.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi>=0.115.12,<1.0.0",
  "passlib>=1.7.4,<2.0.0",
  "PyJWT>=2.10.1,<3.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "redis>=5.2.1,<6.0.0",
  "sqlalchemy>=2.0.39,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0",
  "ruff>=0.11.2,<1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
